modularized the tool, and added remote acquisition
modularized the tool into modules and added remote acquisition to the memory_acquisition.py
