+++ vt_accounts.json
@@ -1,102 +1,82 @@
 [
-    {
-        "email":"cohim16735@owube.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"7f12eae1309627e7365067d54e49e8b691440b9fcc5f8f1de9c7c41a153769a6"
-    },
-    {
-        "email":"segeniv594@ronete.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"fc284ed190940975451c7a1a449073fe5eda3cbd25e572ba1ef086567c71023b"
-    },
-    {
-        "email":"megamap782@kelenson.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"2ab4257bb33830f7a7b0eeac46306251c3e2a525b8c0f4d53cc680469ad1e1c9"
-    },
-    {
-        "email":"sojed67307@chosenx.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"d2003a73eb24b91024a5dc3550d2b9ebeb0580c0852f60fa8a5afd69666d3f7f"
-    },
-    {
-        "email":"xotaf62471@owube.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"caa95d963a84bb0dfcc245247a2b33e17a08edbe9c78fa66b1b46ddb562f66c0"
-    },
-    {
-        "email":"temifi7757@cctoolz.com",
-        "password":"P@ssw0rd",
-        "APIKEY":"1f9b5f7f8683a740e330370009872262bb7659913b3a809ce3d4720188f580c0"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"4b22d28ee5ee3dc6ba8d7442e2d118f83e3d328a0fcbbb03e153d7a769ed3953"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"fa7e9d46b8789f19cbd78cb4da310c4c45c3bf3a57c3bba81b654e76021ef6b0"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"a134f9c98baedf7e05f24229a65b15c4582eb928bdc9fe80ea7994f47ecad30d"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"51f8e18b982a9c010f622a4bfba4d832abc3b61e4d65b23f03108771f5dd2796"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"872014500d7083de6c2a747986e7981a95268c94ca4c233dac68a590a0a4d74b"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"9eebe3335ca174f7385e65d8ab4f9670c68bcfc826a54ebed0d283688658a78b"
-    },
-    {
-        "email":"",
-        "password":"",
-        "APIKEY":"27cd509ce7bf9ff21c38d0c7c0fd3ccf84111c4d424c92ffc3e6e6e8b798aa07"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"5f301dc21f93bcb8e0b7f4d6e6d7b4b6472066d6c90fea51fab6013da5c29117"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"2ac0700a09a9c9234b3629403c61e4d5574d095bbeb37d0c8c5874cbe60f9e31"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"f92d76e64fa94e7ce581230acc4fd5d1eff1ff6efe97a427c48cfdad8732abf5"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"cb6ed3841662a8d9ef5f876b7c70a5d4b926a5a52cb9331b0e674003f5284652"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"1256a751b62c27ffa6841850b8c96c5bb75c52b97a6a7781b9eae8e1dc4b67a6"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"33d546c4ad40109be2aeaad8c094ebcbd48ba5471ca7d46237bb5956d14e22d4"
-    },
-    {
-        "email":"",
-        "password":"P@ssw0rd",
-        "APIKEY":"9f409be26dd5c3e90ac5dbe775c23bc8a9f5336e6690ca10b969aa803b3b2ea0"
-    }
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "5f301dc21f93bcb8e0b7f4d6e6d7b4b6472066d6c90fea51fab6013da5c29117"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "2ac0700a09a9c9234b3629403c61e4d5574d095bbeb37d0c8c5874cbe60f9e31"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "f92d76e64fa94e7ce581230acc4fd5d1eff1ff6efe97a427c48cfdad8732abf5"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "cb6ed3841662a8d9ef5f876b7c70a5d4b926a5a52cb9331b0e674003f5284652"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "1256a751b62c27ffa6841850b8c96c5bb75c52b97a6a7781b9eae8e1dc4b67a6"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "33d546c4ad40109be2aeaad8c094ebcbd48ba5471ca7d46237bb5956d14e22d4"
+  },
+  {
+    "email": "",
+    "password": "P@ssw0rd",
+    "APIKEY": "9f409be26dd5c3e90ac5dbe775c23bc8a9f5336e6690ca10b969aa803b3b2ea0"
+  },
+  {
+    "email": "megamap782@kelenson.com",
+    "password": "P@ssw0rd",
+    "APIKEY": "2ab4257bb33830f7a7b0eeac46306251c3e2a525b8c0f4d53cc680469ad1e1c9"
+  },
+  {
+    "email": "sojed67307@chosenx.com",
+    "password": "P@ssw0rd",
+    "APIKEY": "d2003a73eb24b91024a5dc3550d2b9ebeb0580c0852f60fa8a5afd69666d3f7f"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "4b22d28ee5ee3dc6ba8d7442e2d118f83e3d328a0fcbbb03e153d7a769ed3953"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "fa7e9d46b8789f19cbd78cb4da310c4c45c3bf3a57c3bba81b654e76021ef6b0"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "a134f9c98baedf7e05f24229a65b15c4582eb928bdc9fe80ea7994f47ecad30d"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "51f8e18b982a9c010f622a4bfba4d832abc3b61e4d65b23f03108771f5dd2796"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "872014500d7083de6c2a747986e7981a95268c94ca4c233dac68a590a0a4d74b"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "9eebe3335ca174f7385e65d8ab4f9670c68bcfc826a54ebed0d283688658a78b"
+  },
+  {
+    "email": "",
+    "password": "",
+    "APIKEY": "27cd509ce7bf9ff21c38d0c7c0fd3ccf84111c4d424c92ffc3e6e6e8b798aa07"
+  }
 ]
